fix(acad): event handling on doc and window change events
