+++ ConnectorAutocadCivil/ConnectorAutocadCivil/Entry/App.cs
@@ -40,12 +40,13 @@ namespace Speckle.ConnectorAutocadCivil.Entry
         //the below should fix it! This affects Avalonia and Material 
         AppDomain.CurrentDomain.AssemblyResolve += new ResolveEventHandler(OnAssemblyResolve);
 
-        // set up bindings and subscribe to doument events
+        // set up bindings and subscribe to document events
         SpeckleAutocadCommand.Bindings = new ConnectorBindingsAutocad();
         SpeckleAutocadCommand.Bindings.SetExecutorAndInit();
 
         // for DUI2
         SpeckleAutocadCommand2.Bindings = new ConnectorBindingsAutocad2();
+        SpeckleAutocadCommand2.Bindings.RegisterAppEvents();
       }
       catch(System.Exception e)
       {
