+++ ConnectorAutocadCivil/ConnectorAutocadCivil/UI/ConnectorBindingsAutocadCivil2.cs
@@ -18,8 +18,6 @@ using Speckle.Core.Transports;
 using Speckle.ConnectorAutocadCivil.Entry;
 using Speckle.ConnectorAutocadCivil.Storage;
 
-using AcadApp = Autodesk.AutoCAD.ApplicationServices;
-using AcadDb = Autodesk.AutoCAD.DatabaseServices;
 using Autodesk.AutoCAD.EditorInput;
 using Autodesk.AutoCAD.ApplicationServices;
 using Autodesk.AutoCAD.Colors;
@@ -32,7 +30,7 @@ namespace Speckle.ConnectorAutocadCivil.UI
 {
   public partial class ConnectorBindingsAutocad2 : ConnectorBindings
   {
-    public Document Doc => Application.DocumentManager.MdiActiveDocument;
+    public static Document Doc => Application.DocumentManager.MdiActiveDocument;
 
     // AutoCAD API should only be called on the main thread.
     // Not doing so results in botched conversions for any that require adding objects to Document model space before modifying (eg adding vertices and faces for meshes)
@@ -684,6 +682,23 @@ namespace Speckle.ConnectorAutocadCivil.UI
     #endregion
 
     #region events
+    public void RegisterAppEvents()
+    {
+      //// GLOBAL EVENT HANDLERS
+      Application.DocumentWindowCollection.DocumentWindowActivated += Application_WindowActivated;
+      Application.DocumentManager.DocumentActivated += Application_DocumentActivated;
+      Doc.BeginDocumentClose += Application_DocumentClosed;
+    }
+
+    //checks whether to refresh the stream list in case the user changes active view and selects a different document
+    private void Application_WindowActivated(object sender, DocumentWindowActivatedEventArgs e)
+    {
+      if (e.DocumentWindow.Document == null)
+        return;
+
+      var streams = GetStreamsInFile();
+      UpdateSavedStreams(streams);
+    }
 
     private void Application_DocumentClosed(object sender, DocumentBeginCloseEventArgs e)
     {
@@ -698,11 +713,15 @@ namespace Speckle.ConnectorAutocadCivil.UI
     private void Application_DocumentActivated(object sender, DocumentCollectionEventArgs e)
     {
       // Triggered when a document window is activated. This will happen automatically if a document is newly created or opened.
+      if (e.Document == null)
+        return;
+
       var streams = GetStreamsInFile();
       if (streams != null && streams.Count != 0)
+      {
         SpeckleAutocadCommand2.CreateOrFocusSpeckle();
-
-      UpdateSavedStreams(streams);
+        UpdateSavedStreams(streams);
+      }
     }
     #endregion
   }
